requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.51.0
importlib_resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
torch==1.8.2+cu111
torchaudio==0.8.2
torchsummary==1.5.1
torchvision==0.9.2+cu111
typing_extensions==4.11.0
zipp==3.18.1
